move tests (and fix crossover more)
